rtcom-log-model: start-time is 64 bit integer

+++ src/rtcom-log-view.c
@@ -921,7 +921,7 @@ _text_cell_func(
   gchar *remote_name;
   gchar *remote_uid;
   gchar *text;
-  gint timestamp;
+  gint64 timestamp;
   gint count;
   gchar *group_title;
   gchar *markup;
